Fixed an important lag issue
This commit fixes an issue in which the client would sometimes see large
cells when very small. This might not seem like a problem, but it has to
do with the way cells are rendered. There are 2 types of rendering
cells:
1. A precise method, each point on the circle is calculated for that
wobbly effect. Used when the client has small cells.
2. Smooth rendering when the client has large cells; a full circle is
rendered, the precise rendering would lag too much as there would be too
many points to render.

The issue this commit fixes would cause the large cells that are on the
map to appear on the client's visible area, forcing the client to render
them, even though the client would not see it on screen. If the client
had small cell(s), there would be a lot of points to render, causing the
lag.